Add multiple process support to startup script.
It's up to the user to modify galaxy.ini and job_conf.xml as necessary
to actually *use* multiple processes. Also note that nginx is configured
only to proxy for the process named 'main' running on port 8080; as
such, this change by itself is really only useful for adding job
handlers.

The server processes *must* be restarted after tool installation in
order for all the processes to see the new tool. This is a known issue.